Configure vagrant NFS synced folder
Not sure why I had to do this unless the Ubuntu box isn't configured
for it by default. With no synced_folder declaration there's no synced
folder even though it's supposed to be enabled by default, and with
no nfs type declaration it's implemented with rsync which isn't that
useful.