Simplify data model setup
We now download all the yaml files in one bundle (that was built by
the CI job). This allows us to use the ansible "command" operator
instead of having to use the "shell" operator so we don't need "warn:
false" anymore.

We use kustomize instead of "kubectl -k" because the version of
kustomize that's baked into kubectl is old and incomplete, so we need
to use the standalone kubectl instead.

We don't create the acnodal namespace in main.yml anymore because the
egw-resource-model manifests create it so we don't have to do it
ad-hoc anymore.