Install a daemon that configures Envoy pod networking
It has to run outside of Kubernetes because we need to scan all of the
linux network namespaces to find the veth pair that's attached to the
Envoy pod.

Since it's running ad-hoc I thought I'd see how well Python works in a
k8s context. Answer: very well indeed! Python is one of the two
fully-supported client languages (Go being the other) so it's got
high-quality client bindings. It's also got bindings for almost
everything else that you can think of, although in our case Docker is
the most important.