Write service pod veth ifindex to the loadbalancer status
We need the ifindex of the Envoy proxy pod's veth to configure tunnels
but we can only find it from outside of k8s because our algorithm
depends on us having root access to all of the network namespaces.

To allow processes running in k8s to use the ifindex the Python daemon
caches it in the loadbalancer CRs status.