Add a shell provisioner to fix some Vagrant gotchas
We want to be able to use the Ansible "egw" role both with Vagrant and
standalone from the command line. This means that we want to prepare
the vagrant-created guest VM so it looks like what we expect. Using a
shell provisioner lets us keep the config changes in the same place as
the other Vagrant configuration.