Only override controller-manager docker image when env var is set
Now that the egw-resource-model CI job sets a precise docker image
label we don't want to have a default value here. The default should
be what the CI job set. Now if the EPIC_CM_IMG env var isn't set we
use the CI-generated deployment manifest as-is, but if it is set we
run kustomize to override the image.