Fix a "make up" flake
Sometimes when we tried to create the EGW config singleton one of the
controller-manager webhooks would not be fully running so the
operation would fail. We were running "wait" but evidently we weren't
waiting for the right things.

This approach is simpler: instead of adding an arbitrary wait, we just
try to do what we want to do. If it fails, we back off and try again.
It also seems to be more robust.