Allow marin3r image override
We need to run a custom version of marin3r, at least for a
while. Hopefully this will just be temporary, but we've had to make
some fixes for Envoy V3 protocol.

Marin3r is a little more complicated than other overrides because we
need to apply the override in two places: the marin3r deployment
manifest (which determines the controller-manager and webhook image),
and the epic singleton (which controls the discoveryservice image).