Don't add an IP address to the multus bridge
We want more flexibility in assigning IP addresses to Envoy pods so we
don't want the CNI IPAM second-guessing the assignments that we
make. Resource model v0.7.0-pre3 runs the IPAM in "static" mode so it
simply returns the IP address that we put in the deployment annotation back
to the bridge plugin. This gives us ultimate flexibility in assigning
IPs to pods.