Install epic resource model using downloaded templates
We added configuration settings to the snap package so we want roughly
equivalent functionality here. We don't prompt since this isn't
intended for users, but we have the same set of config params in
epic.yml that we used to prompt for.

Now we download a package of deployment manifest templates and use
Ansible's Jinja2 templating to process them before applying.