Each Gateway rule match becomes its own Envoy route
We need to "unroll" the HTTPRoutes more than we were. We had each
route rule as an Envoy route, but then we could end up with a single
rule that had two "match" keys. Now we iterate over Matches at a
higher level so a rule with two matches will turn into two Envoy
routes.

It's not completely there yet but it passes more conformance tests
than it did before, and in general provides a more sensible Envoy
configuration.