Use the pure-python protobuf driver
One of the setup-network daemon's dependencies upgraded to a new major
version that no longer worked, but luckily they indicated in the error
message that this is a workaround. The pure-python driver is probably
slower than the native one but we're doing so few transactions with
the control plane that it won't make a measurable difference.