Implement Gateway redirect filters
There are two families of Gateway rule filters. One causes the request
to be modified before it is proxied upstream. The other causes Envoy
to respond to the request itself so it's never forwarded.

We had implemented most of the first family but not the second. Now we
can do things like redirect all HTTP requests to their corresponding
HTTPS endpoint.