Install apiserver on the default port
There's a regression in kubeadm where if you specify a custom
apiServerEndpoint port only the server gets configured, not clients,
so kubelet can't find the apiserver. For now we'll switch back to the
default port and if we need to run on non-default we'll have to figure
out what the story is.

Signed-off-by: Toby Cabot <toby@acnodal.io>