Fix adding nodes to clusters
We can now add nodes to existing clusters, either epic or mk8s. By
default, epic1 is in the "epic" group which Ansible will configure as
a full-blown epic k8s cluster. epic2 and epic3 are in the "node" group
and will join the existing epic1 cluster.

The same "node" playbook works for both types of cluster, we just need
to tell it which script to run to join: either the epic one or the
client one.

Signed-off-by: Toby Cabot <toby@acnodal.io>