Don't prompt the user for the "become" credentials
We now store the become creds (i.e., for ssh and sudo) in the vault
file and use command-line params to tell Ansible how to find it. This
lets us store the password semi-secretly and not have to type it every
time. The default values are what Vagrant uses.

Since the vault is now used for every command invocation (not just the
epic playbook) I moved it up to the project root.

Signed-off-by: Toby Cabot <toby@acnodal.io>