Simplify the default config by eliminating the internal bridge
This added some complexity to the playbook because we need to figure
out what the server's dhcp-assigned address is, but it results in a
much more streamlined Vagrant setup.

We used to have an internal bridge for the VMs to talk to one another,
but this is easier to work with. We install the eBPF programs on eth0
and use the Vagrant management network as-is.

Signed-off-by: Toby Cabot <toby@acnodal.io>