Upgrade to Gateway API v1.0.0
We're now using v1.0.0 of the Gateway CRD. This means that users can
no longer use the gateway.networking.k8s.io/v1alpha2 apiVersion for
all objects. It's a little strange now, some of the CRDs are still in
v1alpha2 but some are in v1beta1, some are in v1, and some are in more
than one apiVersion.